Extract object changes implementation classes from interfaces file
