+++ packages/library-base/src/commonMain/kotlin/io/realm/notifications/ObjectChange.kt
@@ -41,24 +41,3 @@ interface UpdatedObject<O : RealmObject> : ObjectChange<O> {
     }
 }
 interface DeletedObject<O : RealmObject> : ObjectChange<O>
-
-internal class InitialObjectImpl<O : RealmObject>(override val obj: O?) : InitialObject<O> {
-    override val state: ObjectChange.State
-        get() = ObjectChange.State.INITIAL
-}
-
-internal class UpdatedObjectImpl<O : RealmObject>(
-    override val obj: O?,
-    override val changedFields: Array<String>
-) : UpdatedObject<O> {
-    override val state: ObjectChange.State
-        get() = ObjectChange.State.UPDATED
-}
-
-internal class DeletedObjectImpl<O : RealmObject> : DeletedObject<O> {
-    override val state: ObjectChange.State
-        get() = ObjectChange.State.DELETED
-
-    override val obj: O?
-        get() = null
-}
